Simplify argument defaulting as sys.modules is a constant binding
Python's argument defaulting is broken by design:

  default_x = 2

  def foo(x=default_x):
      ...

will cause x to have the default value of 2 in foo, even if you change
default_x: argument defaults are evaluated at definition time when
they should be evaluated at call time.  So you normally need this horror:

  def foo(x=None):
      if x is None:
          x = default_x

But in the case where the binding of the default is *constant* you can
get away with the more natural thing.

I was doing the hairy thing to default arguments to sys.modules, but
the binding of sys.modules is (or should be) a constant.

So this commit changes that.